Moved Analytics file to shared folder
